Refactoring Daemon code
